feature/Post-Management: Finished posting jobs based on active date feature

+++ JobApp/src/main/java/com/ubt/andi/jobapp/controllers/ApplicationsController.java
@@ -28,6 +28,7 @@ public class ApplicationsController {
     @GetMapping("/job/{jobId}/apply")
     public String getApplicationForm(@PathVariable("jobId") Long jobId,Model model){
         Job jobDb = jobService.getJobByIdAndUser(jobId);
+        if(!jobDb.isActive()) return "redirect:/";
         AppUser user = userService.findUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
         Application existingApplication = applicationService.findApplicationByUserAndJob(user, jobDb);
         if(existingApplication == null){
@@ -42,6 +43,7 @@ public class ApplicationsController {
     @PostMapping("/job/{jobId}/apply")
     public String sendApplication(@PathVariable("jobId") Long jobId, @RequestParam("cvUpload") MultipartFile cvUpload, @ModelAttribute("application") Application application){
         Job jobDb = jobService.getJobByIdAndUser(jobId);
+        if(!jobDb.isActive()) return "redirect:/";
         AppUser user = userService.findUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
         Application existingApplication = applicationService.findApplicationByUserAndJob(user, jobDb);
         if(existingApplication != null){
@@ -57,6 +59,7 @@ public class ApplicationsController {
     public String deleteApplication(@PathVariable("jobId") Long jobId){
         AppUser userDb = userService.findUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
         Job jobDb = jobService.getJobByIdAndUser(jobId);
+        if(!jobDb.isActive()) return "redirect:/";
         Application applicationDb = applicationService.findApplicationByUserAndJob(userDb,jobDb);
         fileUploadService.deleteDocument(applicationDb);
         applicationService.deleteApplication(applicationDb);
